Correction of the contract's mistakes
