--- BoulderDash-BaseProject/BoulderDash/contract/Shared/IBoulderDashController.java
@@ -1,6 +1,0 @@
-package Shared;
-
-public interface IBoulderDashController {
-	public void play();
-	public IOrderPerformer getOrderPerformer();
-}
