--- BoulderDash-BaseProject/BoulderDash/contract/Shared/IBoulderDashModel.java
@@ -1,6 +1,0 @@
-package Shared;
-
-public interface IBoulderDashModel {
-	public IMap getMap();
-	public IMobile getRockford();
-}
