--- BoulderDash-BaseProject/BoulderDash/contract/Shared/IBoulderDashView.java
@@ -1,6 +1,0 @@
-package Shared;
-
-public interface IBoulderDashView {
-	public void displayMessage(String message);
-	public void followRockford();
-}
