--- BoulderDash-BaseProject/BoulderDash/contract/Shared/ICAD.java
@@ -1,8 +1,0 @@
-package Shared;
-
-public interface ICAD {
-	public void connect();
-	public void close();
-	public char[][]getMap(int idMap);
-	public void saveNewMap();
-}
