--- BoulderDash-BaseProject/BoulderDash/contract/Shared/IElement.java
@@ -1,7 +1,0 @@
-package Shared;
-
-public interface IElement {
-	public void getSprite();
-	public void getPermeability();
-	public void getImage();
-}
