--- BoulderDash-BaseProject/BoulderDash/contract/Shared/IEnnemy.java
@@ -1,5 +1,0 @@
-package Shared;
-
-public interface IEnnemy {
-	public void loot();
-}
