--- BoulderDash-BaseProject/BoulderDash/contract/Shared/IMap.java
@@ -1,11 +1,0 @@
-package Shared;
-
-import java.util.Observable;
-
-public interface IMap {
-	public int getWidth();
-	public int getHeight();
-	public IElement getOnTheRoad(int x, int y);
-	public void setMobileHasChanged();
-	public Observable getObservable();
-}
