--- BoulderDash-BaseProject/BoulderDash/contract/Shared/IOrderPerformer.java
@@ -1,5 +1,0 @@
-package Shared;
-
-public interface IOrderPerformer {
-	public void orderPerformer(UserOrder userOrder);
-}
