--- BoulderDash-BaseProject/BoulderDash/contract/Shared/Permeability.java
@@ -1,6 +1,0 @@
-package Shared;
-
-public enum Permeability {
-	BLOCKING,
-	PENETRABLE
-}
