--- BoulderDash-BaseProject/BoulderDash/contract/Shared/UserOrder.java
@@ -1,9 +1,0 @@
-package Shared;
-
-public enum UserOrder {
-	UP,
-	DOWN,
-	RIGHT,
-	LEFT,
-	NOP
-}
