+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/common/Permeability.java
@@ -1,0 +1,6 @@
+package common;
+
+public enum Permeability {
+	BLOCKING,
+	PENETRABLE
+}
