+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/common/Sprite.java
@@ -1,0 +1,58 @@
+package common;
+
+import java.awt.Image;
+
+public class Sprite {
+	
+	private Image image;
+	private String imageName;
+	private char consoleImage;
+	private Boolean imageLoaded;
+	
+	public Sprite(char character){
+		
+	}
+	
+	public Sprite(char character, String imageName){
+		
+	}
+
+	public Image getImage() {
+		return image;
+	}
+
+	public void setImage(Image image) {
+		this.image = image;
+	}
+
+	public String getImageName() {
+		return imageName;
+	}
+
+	public void setImageName(String imageName) {
+		this.imageName = imageName;
+	}
+
+	public char getConsoleImage() {
+		return consoleImage;
+	}
+
+	public void setConsoleImage(char consoleImage) {
+		this.consoleImage = consoleImage;
+	}
+
+	public void setImageLoaded(Boolean imageLoaded) {
+		this.imageLoaded = imageLoaded;
+	}
+	
+	public Boolean isNameLoaded(){
+		return true;
+	}
+	
+	public void loadedImage(){
+		
+	}
+	
+	
+	
+}
