+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/common/UserOrder.java
@@ -1,0 +1,9 @@
+package common;
+
+public enum UserOrder {
+	UP,
+	DOWN,
+	RIGHT,
+	LEFT,
+	NOP
+}
