+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/controller/IBoulderDashController.java
@@ -1,0 +1,6 @@
+package controller;
+
+public interface IBoulderDashController {
+	public void play();
+	public IOrderPerformer getOrderPerformer();
+}
