+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/controller/IOrderPerformer.java
@@ -1,0 +1,7 @@
+package controller;
+
+import common.UserOrder;
+
+public interface IOrderPerformer {
+	public void orderPerformer(UserOrder userOrder);
+}
