+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/model/IBoulderDashModel.java
@@ -1,0 +1,6 @@
+package model;
+
+public interface IBoulderDashModel {
+	public IMap getMap();
+	public IMobile getRockford();
+}
