+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/model/ICAD.java
@@ -1,0 +1,8 @@
+package model;
+
+public interface ICAD {
+	public void connect();
+	public void close();
+	public char[][]getMap(int idMap);
+	public void saveNewMap();
+}
