+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/model/IElement.java
@@ -1,0 +1,7 @@
+package model;
+
+public interface IElement {
+	public void getSprite();
+	public void getPermeability();
+	public void getImage();
+}
