+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/model/IEnnemy.java
@@ -1,0 +1,5 @@
+package model;
+
+public interface IEnnemy {
+	public void loot();
+}
