+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/model/IMap.java
@@ -1,0 +1,11 @@
+package model;
+
+import java.util.Observable;
+
+public interface IMap {
+	public int getWidth();
+	public int getHeight();
+	public IElement getOnTheRoad(int x, int y);
+	public void setMobileHasChanged();
+	public Observable getObservable();
+}
