+++ BoulderDash-BaseProject/BoulderDash/contract/src/main/java/view/IBoulderDashView.java
@@ -1,0 +1,6 @@
+package view;
+
+public interface IBoulderDashView {
+	public void displayMessage(String message);
+	public void followRockford();
+}
